+++ typescript/action_kit_api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.0.0
+
+ - **Breaking:** The type `ParameterOption` was renamed to `ExplicitParameterOption`.
+ - Ability to auto-generate supported options through the `ParameterOption` type `ParameterOptionsFromTargetAttribute`.
+ - Support for metric queries and responses.
+
 ## 1.0.0
 
  - Initial release